updated organizer create competition route; fixed a util function

+++ util/type.go
@@ -3,7 +3,7 @@ package util
 func InterfaceSliceToIntSlice(input []interface{}) []int {
 	output := make([]int, 0)
 	for _, each := range input {
-		output = append(output, int(each.(float64)))
+		output = append(output, int(each.(int)))
 	}
 	return output
 }
